Default template of a schema description

Default template of a schema description

schema_default

Format

The object of class schema describes at which position in a table which information can be found. It contains the two obligatory slots clusters and variables.

There is hardly any limit to how data can be arranged in a spreadsheet, apart from the apparent organisation into a lattice of cells. However, it is mostly the case that data are gathered in areas of a spreadsheet that can be outlined by a rectangle. Those rectangles, including more or less coherent tables, are what is considered 'cluster' in arealDB. Clusters are described by the properties:

top [integerish(1)]:
The vertical cell values of the top-left cell of each cluster. This can also be a vector of values in case there are several clusters.

left [integerish(1)]:
The horizontal cell value of the top-left cell of each cluster. This can also be a vector of values in case there are several clusters.

width [integerish(1)]:
The width of each cluster in cells. This can also be a vector of values in case there are several clusters

height [integerish(1)]:
The height of each cluster in cells. This can also be a vector of values in case there are several clusters

id [character(1)]:
It is typically the case that data are clustered according to one of the variables of interest. In such cases, this variable needs to be registered here.

header [logical(1)]:
In cases where a cluster contains a table in which variables are nested, i.e., where column names are in more than one row, header must be set to FALSE.

Each element in the slot variables is a list that describes one of the variables that shall be comprised in the final database. Variables are either so-called identifying (or ID) variables, which indicate observation units, or values variables, which carry the observed values. Identifying variables are described by the properties:

type [character(1)]:
The value "id" signals that this is an identifying variable.

name [character(1)]:
A name that would be assigned as column name in the database instead of the list name.

split [character(1)]:
A regular expression that matches the part of values that are supposed to be part of the variable, when the cells contain more than one value (for example seperated by a ",").

row [integerish(1)]:
The row(s) in which the variable values are recorded.

col [integerish(1)]:
The column(s) in which the variable values are recorded.

rel [logical(1)]:
Whether or not the values in row and col are relative to the cluster positions or whether they refer to the overall table.

Values variables are described by the properties:

type [character(1)]:
the value "values" signals that this is a values variable.

unit [character(1)]:
The unit in which the values shall be recorded in the database.

factor [character(1)]:
A factor to transform the values to unit. For instance, if values are recorded in acres, but shall be contained in the database in hectare, the factor would be 0.40468.

row [integerish(1)]:
The row(s) in which the variable values are recorded.

col [integerish(1)]:
The column(s) in which the variable values are recorded.

rel [logical(1)]:
Whether or not the values in row and col are relative to the cluster positions or whether they refer to the overall table.

key [character(1)]:
If the variable is recorded, together with other variables, so that the variable names are listed in one column and the respective values are listed in another column, give here the name of the column that contains the variable names.

value [character(1)]:
If the variable is recorded, together with other variables, so that the variable names are listed in one column and the respective values are listed in another column, give here the level in the names column that stands for the values of this variable.

The default schema description contains an example of an identifying and a values variables. Further identifying and values variables would be added when more variables are contained in a table.